util: fix `Setsocksize()` which was a noop most of the time.
`getsockopt()` should be called with a pointer to the size of the
before-last argument as last argument. Since the `len` argument was
unitialized, it could be assigned to some register and initialized to
0. If this was not the case for `saved`, it was initialized to some
random value. Hence the `value > saved` value was likely to be false
and `Setsocksize()` didn't do anything.